% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_abundance_plot.R
\name{aa_abundance_plot}
\alias{aa_abundance_plot}
\title{Plot abundance of unique amino acids in sequence}
\usage{
aa_abundance_plot(aa_seq)
}
\arguments{
\item{aa_seq}{string of amino acids}
}
\description{
Plot abundance of unique amino acids in sequence
}
