% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNA_to_RNA.R
\name{DNA_to_RNA}
\alias{DNA_to_RNA}
\title{DNA to RNA sequence transcription function}
\usage{
DNA_to_RNA(DNA)
}
\arguments{
\item{DNA}{sequence to be transcribed}
}
\value{

}
\description{
DNA to RNA sequence transcription function
}
