% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RNA}
\alias{RNA}
\title{RNA codons}
\format{

}
\source{
\url{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=tgencodes#SG1}
}
\usage{
RNA
}
\description{
Translation codons from RNA to AA
}
\keyword{datasets}
