% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_transform.R
\name{rna_transform}
\alias{rna_transform}
\title{Substring RNA in a window of 3 letters}
\usage{
rna_transform(rna_seq, start = 1)
}
\arguments{
\item{rna_seq}{A character vector.}

\item{start}{integer. The first element to begin codon transformation.}
}
\description{
Substring RNA in a window of 3 letters
}
