% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Translate RNA codons to amino acids}
\usage{
translate(codons)
}
\arguments{
\item{codons}{}
}
\value{
amino acids sequence
}
\description{
Translate RNA codons to amino acids
}
